requirements.txt
langchain
openai
pinecone-client
transformers
datasets
torch
faiss-cpu
tiktoken
unstructured
langchain_community
pymupdf
sentence-transformers
openai
